% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{CombIndex1D}
\alias{CombIndex1D}
\title{Combined Index with Holes and Skewness Index}
\usage{
CombIndex1D(origdata,
                   proj,
                   gamma=0.9)
}
\arguments{
\item{origdata}{the original data matrix}

\item{proj}{projection vector you want to know the index value}

\item{gamma}{ratio for combining two indices}
}
\value{
numeric index value
}
\description{
Find interesting patterns from data
}
\examples{
 \donttest{
data(sampledata)
CombIndex1D(sampledata[,-1],c(1,0,0,0))
}

}
