% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuberplotMod.R
\name{HuberplotMod}
\alias{HuberplotMod}
\title{Huber plot}
\usage{
HuberplotMod(origdata, 
                    group=NULL, 
                    PPmethod = "Holes", 
                    opt.proj = TRUE, 
                    UserDefFtn = NULL,
                    plotFlag = TRUE,
                    main=NULL,...)
}
\arguments{
\item{origdata}{2-dimensional numerical data for Huber plot}

\item{group}{class information vector of data}

\item{PPmethod}{method for projection pursuit; "Holes", "Skew", "Comb", 
"NaturalHermite", "Lpp", "P","UserDefProjData", and "UserDefData"}

\item{opt.proj}{flag to show the best projection in the plot}

\item{UserDefFtn}{User defined index function when PPmethod="UserDefProjData" or 
"UserDefData"}

\item{plotFlag}{flag to show huberplot}

\item{main}{Main title of the plot}

\item{...}{Arguments to be passed to methods.}
}
\value{
huber plot
}
\description{
Huber plot for 2D data. Plot index values according to 360 
degree angle, and histogram of optimally projected data. Modified for the 
projection pursuit index without group information.
}
\examples{
 \donttest{
 data(sampledata)
HuberplotMod(origdata = sampledata[,c(2,3)],
             group = sampledata$species,
             PPmethod="Holes",
             main=paste("Holes Index"))  
             
HuberplotMod(origdata = sampledata[,2:3],
             PPmethod="UserDefData",
             UserDefFtn=HolesIndex1D,
             main=paste("Holes Index"))
}


}
\references{
...
}
