% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{LppIndex1D}
\alias{LppIndex1D}
\title{Lpp Index (Espezua et al., 2015)}
\usage{
LppIndex1D(origdata, 
                  proj, 
                  method="kNN",
                  eps=0,
                  prob=0.2,
                  weight=FALSE,
                  nk=0,
                  t=1)
}
\arguments{
\item{origdata}{the original data matrix}

\item{proj}{projection vector you want to know the index value}

\item{method}{"kNN" determines neighbors based on count, while "dist" 
determines based on distance.}

\item{eps}{distance criterion at the "dist" method}

\item{prob}{ratio for determining the neighborhood for kNN method}

\item{weight}{If "TRUE", the target function use a weight based on distance; 
if "FALSE", no weight}

\item{nk}{count criterion at the "kNN" method}

\item{t}{weight parameter}
}
\value{
numeric index value
}
\description{
Find projections in which close points on the original 
              data are also located close after projection
}
\examples{
 \donttest{
 data(sampledata)
 LppIndex1D(sampledata[,-1],c(1,0,0,0))
}

}
\references{
Espezua, S., Villanueva, E., Maciel, C. D., & Carvalho, A. 
(2015). A Projection Pursuit framework for supervised dimension reduction of 
high dimensional small sample datasets. Neurocomputing, 149, 767-776.

He, X., & Niyogi, P. (2003). Locality preserving projections. 
Advances in neural information processing systems, 16.
}
