% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{NHIndex1D}
\alias{NHIndex1D}
\title{Natural Hermite Index}
\usage{
NHIndex1D(origdata,
                 proj,
                 m=0)
}
\arguments{
\item{origdata}{the original data matrix}

\item{proj}{projection vector you want to know the index value}

\item{m}{order of the index. low order captures the overall structure 
of data, high order finds detailed patterns of data}
}
\value{
numeric index value
}
\description{
Fine projections away from normality
}
\examples{
 \donttest{
 data(sampledata)
 NHIndex1D(sampledata[,-1],c(1,0,0,0))
}

}
\references{
Cook, D., Buja, A., & Cabrera, J. (1993). Projection pursuit 
indexes based on orthonormal function expansions. Journal of Computational 
and Graphical Statistics, 2(3), 225-250.
}
