% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{PIndex1D}
\alias{PIndex1D}
\title{P Index (Friedman et al., 1974)}
\usage{
PIndex1D(origdata,
                proj,
                probT=0.05,
                R=0.1)
}
\arguments{
\item{origdata}{the original data matrix}

\item{proj}{projection vector you want to know the index value}

\item{probT}{the ratio to trim data}

\item{R}{criterion for determining neighbors}
}
\value{
numeric index value
}
\description{
Find a direction that data points are gathered inside the 
              clusters, at the same time, separating the clusters
}
\examples{
 \donttest{
data(sampledata)
PIndex1D(sampledata[,-1],c(1,0,0,0))
}

}
\references{
Friedman, J. H., & Tukey, J. W. (1974). A projection pursuit 
algorithm for exploratory data analysis. IEEE Transactions on computers, 
100(9), 881-890.
}
