% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPtreeCluster.R
\name{plot.PPtreecluster}
\alias{plot.PPtreecluster}
\title{Projection Pursuit Cluster Tree Plot}
\usage{
\method{plot}{PPtreecluster}(
  x,
  font.size = 17,
  width.size = 1,
  main = "Projection Pursuit Classification Tree",
  sub = NULL,
  ...
)
}
\arguments{
\item{x}{PPtreecluster object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node}

\item{main}{main title}

\item{sub}{sub title}

\item{...}{arguments to be passed to methods}
}
\description{
Draw projection pursuit cluster tree with tree structure. It is
modified from a function in party library.
}
