% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclustery.R
\name{PPclustery}
\alias{PPclustery}
\title{Find Clusters Using PPtreecluster Object}
\usage{
PPclustery(Tree.result,
                  test.data=NULL,
                  true.class=NULL,...)
}
\arguments{
\item{Tree.result}{PPtreecluster object}

\item{test.data}{the test dataset}

\item{true.class}{true class of test dataset if available}

\item{...}{arguments to be passed to methods}
}
\value{
cluster predicted cluster

true.class true class of test dataset from input parameter

confusion_matrix confusion matrix with cluster and true.class
}
\description{
Find cluster for the test set with the fitted projection pursuit 
cluster tree
}
\examples{
data(sampledata)
n <- nrow(sampledata)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
Cluster.result <- PPTreeCluster(sampledata[train,-1])
PPclustery(Cluster.result,sampledata[test,-1],sampledata[test,1])
}
\keyword{clustering}
\keyword{tree}
