% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclustNodeViz.R
\name{PPclustNodeViz}
\alias{PPclustNodeViz}
\title{PPtreecluster Node Visualization}
\usage{
PPclustNodeViz(PPclusterOBJ,
                      node.id, 
                      TrueClass=NULL,
                      legend=TRUE,
                      std=TRUE,
                      image=FALSE,
                      diffProp=0.2)
}
\arguments{
\item{PPclusterOBJ}{PPtreecluster object}

\item{node.id}{node ID}

\item{TrueClass}{true class information}

\item{legend}{flag to represent legend in the plot. Default value is TRUE}

\item{std}{flag to standardize data before drawing plot}

\item{image}{flag to draw image plot of correlation matrix}

\item{diffProp}{percentage of number of variables with significant 
differences and shown in red in the bar chart style means}
}
\description{
Visualization tools to explore each node of PPtreecluster. 
For the inner node, four plots are provided - the bar chart style plot with
projection pursuit coefficients of each variable, the histogram of the
projected data, the bar chart style plots with means of each variables for 
the left and the right group, and the image plot of correlations.
}
\examples{
data(sampledata)
Cluster.result <- PPTreeCluster(sampledata[,-1])
Cluster.result
PPclustNodeViz(Cluster.result,1)

}
\keyword{tree}
