% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{PPclustopt1D}
\alias{PPclustopt1D}
\title{Optimization for Projection Pursuit Indices}
\usage{
PPclustopt1D(origdata,  
                    PPmethod = "Holes", 
                    gamma=0.9, 
                    method="kNN",
                    eps=0, 
                    prob=0.2,
                    weight=FALSE,
                    nk=0L,
                    t=1,
                    m=0L,
                    probT=0.05, 
                    R=0.1, 
                    energy = 0, 
                    cooling = 0.999,
                    TOL = 0.0001, 
                    maxiter = 50000L)
}
\arguments{
\item{origdata}{data matrix without class information}

\item{PPmethod}{projection pursuit method for clustering; "Holes", "Skew", 
"Comb", "Lpp", "NatrualHermite", and "P"}

\item{gamma}{combine ratio}

\item{method}{"kNN" determines neighbors based on count, while "dist" 
determines based on distance.}

\item{eps}{distance criterion at the "dist" method}

\item{prob}{ratio for determining the neighborhood for kNN method}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{nk}{count criterion at the "kNN" method}

\item{t}{weight parameter}

\item{m}{order of the index. low order captures the overall structure 
of data, high order finds detailed patterns of data}

\item{probT}{the ratio to trim data}

\item{R}{criterion for determining neighbors}

\item{energy}{energy parameter}

\item{cooling}{cooling parameter}

\item{TOL}{tolerance}

\item{maxiter}{number of maximum iteration}
}
\value{
indexbest maximum projection pursuit index value

projbest optimal 1-dim projection matrix

origdata  original data matrix

proj.data projected data onto the best projection
}
\description{
Find the 1-dim optimal projection using various projection pursuit indices 
with class information
}
\examples{
data(sampledata)
PP.proj.result <- PPclustopt1D(sampledata[,-1])
hist(PP.proj.result$proj.data)
}
\keyword{projection}
\keyword{pursuit}
