% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclustoptViz.R
\name{PPclustoptViz}
\alias{PPclustoptViz}
\title{Visualize PPclusteropt Object}
\usage{
PPclustoptViz(PPclustoptOBJ,
                     classInfo=NULL)
}
\arguments{
\item{PPclustoptOBJ}{PPclustoptOBJ object. result from PPclustopt1D}

\item{classInfo}{class information}
}
\description{
Visualize the result of projection pursuit optimization
}
\examples{
data(sampledata)
PP.proj.result <- PPclustopt1D(sampledata[,-1])
PPclustoptViz(PP.proj.result)
PPclustoptViz(PP.proj.result,classInfo=sampledata[,1])
}
