% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTreeCluster.R
\name{PPTreeCluster}
\alias{PPTreeCluster}
\title{Construct The Projection Pursuit Cluster Tree}
\usage{
PPTreeCluster(origdata,
                     PPmethod="Holes",
                     gamma=0.5,
                     method="kNN",
                     eps=0, 
                     prob=0.2,
                     weight=FALSE,
                     nk=0L,
                     t=1,
                     m=0L,
                     probT=0.05, 
                     R=0.1, 
                     max_depth=5,
                     minNodeN=5,
                     energy=0,
                     cooling=0.999,
                     TOL=1/1000000,
                     maxiter=50000,
                     seed=NULL,
                     notenough.plot = FALSE,
                     relCUTOFF=0.1,
                     cutB=0.2, 
                     minNrate=0.1,
                     ...)
}
\arguments{
\item{origdata}{data matrix without class information}

\item{PPmethod}{projection pursuit method for clustering; "Holes", "Skew", 
"Comb", "Lpp", "NatrualHermite", and "P"}

\item{gamma}{combine ratio}

\item{method}{"kNN" determines neighbors based on count, while "dist" 
determines based on distance.}

\item{eps}{distance criterion at the "dist" method}

\item{prob}{ratio for determining the neighborhood for kNN method}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{nk}{count criterion at the "kNN" method}

\item{t}{weight parameter}

\item{m}{order of the index. low order captures the overall structure 
of data, high order finds detailed patterns of data}

\item{probT}{the ratio to trim data}

\item{R}{criterion for determining neighbors}

\item{max_depth}{maximum depth of the projection pursuit cluster tree}

\item{minNodeN}{minimum number of observations in the final node}

\item{energy}{parameter for optimization}

\item{cooling}{parameter for optimization}

\item{TOL}{parameter for optimization}

\item{maxiter}{parameter for optimization}

\item{seed}{seed for random number}

\item{notenough.plot}{flag to draw plot}

\item{relCUTOFF}{parameter for density estimation}

\item{cutB}{parameter to define cluster}

\item{minNrate}{parameter to define cluster}

\item{...}{Arguments to be passed to methods.}
}
\value{
PPtreecluster object
}
\description{
Find tree structure using various projection pursuit indices for
clustering in each split.
}
\examples{
data(sampledata)
Cluster.result <- PPTreeCluster(sampledata[,-1])
Cluster.result
}
