% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPPtreeCluster.R
\name{print.PPtreecluster}
\alias{print.PPtreecluster}
\title{Print PPtreecluster Result}
\usage{
\method{print}{PPtreecluster}(x, coef.print = FALSE, cutoff.print = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{PPTreecluster Importance object}

\item{coef.print}{print projection coefficients in each node if TRUE}

\item{cutoff.print}{print cutoff values in each node if TRUE}

\item{verbose}{print if TRUE, no output if FALSE}

\item{...}{arguments to be passed to methods}
}
\description{
Print the projection pursuit classification tree result
}
