\name{sampledata}
\alias{sampledata}
\docType{data}
\title{
sample data selected from Palmer Archipelago (Antarctica) Penguin Data
}
\description{
The selected data from Palmer Archipelago (Antarctica) Penguin Data without missing.
}

\arguments{
    \item{species}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo).}
    \item{bill_length_mm}{a number denoting bill length (millimeters).}
    \item{bill_depth_mm}{a number denoting bill depth (millimeters).}
    \item{flipper_length_mm}{an integer denoting flipper length (millimeters).}
    \item{body_mass_g}{an integer denoting body mass (grams).}    
}

\examples{
data(sampledata)
dim(sampledata)
}
\keyword{datasets}
