% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp-add.R
\name{SkewIndex1D}
\alias{SkewIndex1D}
\title{Skewness Index Derived From Natural Hermite Index}
\usage{
SkewIndex1D(origdata,
                   proj)
}
\arguments{
\item{origdata}{the original data matrix}

\item{proj}{projection vector you want to know the index value}
}
\value{
numeric index value
}
\description{
Find skewed distribution
}
\examples{
 \donttest{
data(sampledata)
SkewIndex1D(sampledata[,-1],c(1,0,0,0))
}
}
\references{
Cook, D., Buja, A., & Cabrera, J. (1993). Projection pursuit 
indexes based on orthonormal function expansions. Journal of Computational 
and Graphical Statistics, 2(3), 225-250.
}
